### 2023-07-12 #5653 @editorialbot
[PRE REVIEW]: Aletheia: an open-source toolbox for steganalysis
**Submitting author:** @daniellerch (<a href="http://orcid.org/0000-0003-2602-672X">Daniel Lerch-Hostalot</a>)
**Repository:** https://github.com/daniellerch/aletheia
**Branch with paper.md** (empty if default branch): master
**Version:** v0.1
