+++ paper/paper.bib
@@ -7,7 +7,8 @@
     volume={14}, 
     number={5}, 
     pages={1181-1193}, 
-    month={May}
+    month={May},
+    doi={10.1109/TIFS.2018.2871749}
 }
 
 @inproceedings{Fridrich:2007:nsF5,
@@ -56,7 +57,8 @@ isbn="978-3-540-31978-8"
     number={1},
     pages={1-13},
     year={2014},
-    publisher={Springer-Verlag}
+    publisher={Springer-Verlag},
+    doi={10.1186/1687-417X-2014-1}
 }
 
 @inproceedings{Ker:2013:real_world,
@@ -82,7 +84,8 @@ isbn="978-3-540-31978-8"
     pages = {222–229},
     numpages = {8},
     location = {Paris, France},
-    series = {IH\&MMSec'19}
+    series = {IH\&MMSec'19},
+    doi={10.1145/3335203.3335738}
 }
 
 @ARTICLE{Megias:2023,
@@ -92,7 +95,8 @@ isbn="978-3-540-31978-8"
   year={2023},
   volume={20},
   number={2},
-  pages={1403-1421}
+  pages={1403-1421},
+  doi={10.1109/tdsc.2022.3154967}
 }
 
 
@@ -104,7 +108,8 @@ isbn="978-3-540-31978-8"
     pages={4206-4210}, 
     month={Oct},
     publisher={IEEE},
-    address={Paris, France}
+    address={Paris, France},
+    doi={10.1109/icip.2014.7025854}
 }
 
 @inproceedings {Provos:2001:outguess,
@@ -129,6 +134,7 @@ note={Accessed on \today}
     publisher={{S}pringer},
     pages={13-26},
     year={2001},
+    doi={10.1007/3-540-45496-9_2}
 }
 
 
@@ -141,6 +147,7 @@ note={Accessed on \today}
     month={Dec},
     publisher={IEEE},
     address={New York, NY, USA},
+    doi={10.1109/WIFS49906.2020.9360897}
 }
 
 @misc{zhang:2019:steganogan,
