+++ paper/paper.bib
@@ -46,7 +46,8 @@ year="2005",
 publisher="Springer Berlin Heidelberg",
 address="Berlin, Heidelberg",
 pages="119--128",
-isbn="978-3-540-31978-8"
+isbn="978-3-540-31978-8",
+doi={10.1007/11552055_12}
 }
 
 @article{Holub:2014:uniward,
