### 2023-07-19 #5653 @danielskatz
👋 @daniellerch - Thanks for working on the third issue below. Once you have a response to the other two, please let me know.

>
> A few things that we should deal with:
>
> 1 - I see a lot of code in your repo in various languages. Can you give me a high-level overview of it and why there are so many different languages? Are there multiple APIs from each, for example? This is something I would expect to see in the README and paper as well - something about what is the software, and how is it used (though I guess this is a level below the README in the repo to some extent.) To find an editor and reviewers, we need to know a bit about what they need to know.
>
> 2 - Your paper is quite short: ~400 words compared to a JOSS max of 1000, so there is room to add some more info about the software.
>
> 3 - You could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
