+++ FOLDERS.md
@@ -1,0 +1,27 @@
+
+## Folder structure
+
+
+Throughout this document, we will explore the folder structure of the Aletheia project, specifically focusing on the significant directories. Each folder will be examined, providing an explanation of its contents and shedding light on its purpose and contribution to the overall structure. By highlighting these important directories, we aim to provide valuable insights into the project's architecture, aiding in efficient code navigation and comprehension. 
+
+
+### [aletheia-models](/aletheia-models/README.md)
+  
+
+
+### aletheia-octave
+
+
+### aletheia-resources
+
+
+### aletheialib
+
+
+### actors and sample_images
+
+
+
+
+
+
