+++ paper/paper.bib
@@ -73,6 +73,23 @@ doi={10.1007/11552055_12}
 } 
 
 
+@software{Lerch-Hostalot:Aletheia,
+   author = {Lerch-Hostalot, Daniel},
+   doi = {10.5281/zenodo.4655945},
+   month = apr,
+   title = {{Aletheia}},
+   version = {v0.1},
+   year = {2021}
+}
+
+@misc{Lerch-Hostalot:Aletheia,
+   author = {Lerch-Hostalot, Daniel},
+   title = {Aletheia},
+   year = {2023},
+   publisher = {GitHub},
+   journal = {GitHub repository},
+   howpublished = {\url{https://github.com/daniellerch/aletheia}},
+}
 
 @inproceedings{Lerch-Hostalot:2019,
     author = {Lerch-Hostalot, D. and Meg\'{\i}as, D.},
