+++ paper/paper.md
@@ -22,99 +22,143 @@ bibliography: paper.bib
 
 # Summary
 
-Steganalysis is the practice of detecting the presence of hidden information 
-within digital media, such as images, audio, or video. It involves analyzing 
-the media for signs of steganography, which is the technique used to conceal 
-information within the carrier file. Steganalysis techniques can include 
-statistical analysis, visual inspection, and machine learning algorithms to 
-uncover the hidden data. The goal of steganalysis is to determine whether a 
-file contains covert information and potentially identify the steganographic 
+Steganalysis is the practice of detecting the presence of hidden information
+within digital media, such as images, audio, or video. It involves analyzing
+the media for signs of steganography, which is the technique used to conceal
+information within the carrier file. Steganalysis techniques can include
+statistical analysis, visual inspection, and machine learning algorithms to
+uncover the hidden data. The goal of steganalysis is to determine whether a
+file contains covert information and potentially identify the steganographic
 method used.
 
-Steganalysis has become increasingly important in the face of rising espionage 
-and stegomalware threats, particularly in the context of data exfiltration. 
-In this scenario, malicious actors leverage steganographic techniques to 
-conceal sensitive data within innocent-looking files, evading traditional 
-security measures. By detecting and analyzing such covert communication 
-channels, steganalysis helps identify and prevent data exfiltration attempts, 
-safeguarding critical information from falling into the wrong hands. 
+Steganalysis has become increasingly important in the face of rising espionage
+and stegomalware threats, particularly in the context of data exfiltration.
+In this scenario, malicious actors leverage steganographic techniques to
+conceal sensitive data within innocent-looking files, evading traditional
+security measures. By detecting and analyzing such covert communication
+channels, steganalysis helps identify and prevent data exfiltration attempts,
+safeguarding critical information from falling into the wrong hands.
 
-In recent years, there has been a significant growth in the interest of 
+In recent years, there has been a significant growth in the interest of
 researchers towards the field of steganalysis. The application of deep learning
 [@Boroumand:2019:SRNet;@Yousfi:2020:alaska2]
-in steganalysis has opened up new avenues for research, 
-leading to improved detection rates and enhanced accuracy. As the field 
-continues to evolve, experts are actively exploring novel architectures and 
-training methodologies to further refine the performance of deep learning-based 
+in steganalysis has opened up new avenues for research,
+leading to improved detection rates and enhanced accuracy. As the field
+continues to evolve, experts are actively exploring novel architectures and
+training methodologies to further refine the performance of deep learning-based
 steganalysis.
 
 
 # Statement of need
 
-Aletheia addresses two main needs. Firstly, it aims to provide 
-specialized analysts with a tool that implements modern steganalysis algorithms, 
-leveraging deep learning techniques. These algorithms are designed to 
-effectively handle even the most advanced steganography techniques. Secondly, 
-Aletheia serves as a valuable tool for researchers by simplifying the process 
-of conducting experiments and comparing methods. It includes simulators for 
-common algorithms 
+Aletheia addresses two main needs. Firstly, it aims to provide
+specialized analysts with a tool that implements modern steganalysis algorithms,
+leveraging deep learning techniques. These algorithms are designed to
[…]
