+++ aletheia-cache/README.md
@@ -1,0 +1,8 @@
+
+
+## Cache files
+
+Here, among other things, we store code files with incompatible licenses 
+downloaded after user confirmation.
+
+
