### 2023-07-25 a91b8a8
MAXSRM and sigma spam PSRM have been deprecated and removed
