+++ aletheialib/options/feaext.py
@@ -6,7 +6,6 @@ doc="\n" \
 "  - srmq1:         Spatial Rich Models with fixed quantization q=1c.\n" \
 "  - scrmq1:        Spatial Color Rich Models with fixed quantization q=1c.\n" \
 "  - gfr:           JPEG steganalysis with 2D Gabor Filters.\n" \
-"  - hill-maxsrm:   Selection-Channel-Aware Spatial Rich Models for HILL." 
 
 
 # {{{ srm
@@ -119,37 +118,6 @@ def dctr():
     sys.exit(0)
 # }}}
 
-# {{{ hill-sigma-spam-psrm
-def hill_sigma_spam_psrm():
-
-    if len(sys.argv)!=4:
-        print(sys.argv[0], "hill-sigma-spam-psrm <image/dir> <output-file>\n")
-        sys.exit(0)
-
-    import aletheialib.feaext
-
-    image_path=sys.argv[2]
-    ofile=sys.argv[3]
-
-    aletheialib.feaext.extract_features(aletheialib.feaext.HILL_sigma_spam_PSRM_extract, image_path, ofile)
-    sys.exit(0)
-# }}}
-
-# {{{ hill-maxsrm
-def hill_maxsrm():
-
-    if len(sys.argv)!=4:
-        print(sys.argv[0], "hill-maxsrm <image/dir> <output-file>\n")
-        sys.exit(0)
-
-    import aletheialib.feaext
-
-    image_path=sys.argv[2]
-    ofile=sys.argv[3]
-
-    aletheialib.feaext.extract_features(aletheialib.feaext.HILL_MAXSRM_extract, image_path, ofile)
-    sys.exit(0)
-# }}}
 
 
 
