### 2023-07-25 a9abf14
Use Octave structural detectors from the external repository
