+++ testing/feaext.sh
@@ -7,14 +7,14 @@ test(){
    echo "Testing $1 / $2 ..."
    ./aletheia.py $1 $2 fea.txt
    wc -l fea.txt
-   rm -f fea.txt
+   rm -f fea.txt fea.txt.label
 }
 
-#test srm sample_images/alaska2 
+test srm sample_images/alaska2 
 test srmq1 sample_images/lena_gs.png
-#test scrmq1 sample_images/alaska2 
-#test gfr sample_images/alaska2jpg
-#test dctr sample_images/alaska2jpg
+test scrmq1 sample_images/alaska2 
+test gfr sample_images/alaska2jpg
+test dctr sample_images/alaska2jpg
 
 
 
