+++ aletheia-octave/maxSRM/s1/s1.cpp
@@ -1,200 +1,0 @@
-/*
- This class computes and contains co-occurrences of ALL 1st-order residuals
- listed in Figure 1 in our journal HUGO paper (version from June 14), 
- including the naming convention.
-
- List of outputted features:
-
- 1a) spam14h
- 1b) spam14v (orthogonal-spam)
- 1c) minmax22v
- 1d) minmax24
- 1e) minmax34v
- 1f) minmax41
- 1g) minmax34
- 1h) minmax48h
- 1i) minmax54
-
- Naming convention:
-
- name = {type}{f}{sigma}{scan}
- type \in {spam, minmax}
- f \in {1,2,3,4,5} number of filters that are "minmaxed"
- sigma \in {1,2,3,4,8} symmetry index
- scan \in {h,v,\emptyset} scan of the cooc matrix (empty = sum of both 
- h and v scans).
-*/
-
-#include "../mat2D.h"
-#include "../submodel.h"
-#include "../config.cpp"
-#include "../s.h"
-
-#include "s1_spam14h.cpp"
-#include "s1_spam14v.cpp"
-#include "s1_minmax22h.cpp"
-#include "s1_minmax22v.cpp"
-#include "s1_minmax24.cpp"
-#include "s1_minmax34.cpp"
-#include "s1_minmax34h.cpp"
-#include "s1_minmax34v.cpp"
-#include "s1_minmax41.cpp"
-#include "s1_minmax48h.cpp"
-#include "s1_minmax48v.cpp"
-#include "s1_minmax54.cpp"
-
-class s1 : s
-{
-public:
-	void CreateKernels()
-	{
-		mat2D<int> *temp;
-		cutEdgesForParityBy = 1;
-		
-		// Right Kernel
-		temp = new mat2D<int>(3, 3);
-		temp->Write(0, 0, 0); temp->Write(0, 1, 0); temp->Write(0, 2, 0);
-		temp->Write(1, 0, 0); temp->Write(1, 1,-1); temp->Write(1, 2, 1);
-		temp->Write(2, 0, 0); temp->Write(2, 1, 0); temp->Write(2, 2, 0);
-		kerR = temp;
-
-		// Left Kernel
-		temp = new mat2D<int>(3, 3);
-		temp->Write(0, 0, 0); temp->Write(0, 1, 0); temp->Write(0, 2, 0);
-		temp->Write(1, 0, 1); temp->Write(1, 1,-1); temp->Write(1, 2, 0);
-		temp->Write(2, 0, 0); temp->Write(2, 1, 0); temp->Write(2, 2, 0);
-		kerL = temp;
-
-		// Up Kernel
-		temp = new mat2D<int>(3, 3);
-		temp->Write(0, 0, 0); temp->Write(0, 1, 1); temp->Write(0, 2, 0);
-		temp->Write(1, 0, 0); temp->Write(1, 1,-1); temp->Write(1, 2, 0);
-		temp->Write(2, 0, 0); temp->Write(2, 1, 0); temp->Write(2, 2, 0);
-		kerU = temp;
-
-		// Down Kernel
-		temp = new mat2D<int>(3, 3);
-		temp->Write(0, 0, 0); temp->Write(0, 1, 0); temp->Write(0, 2, 0);
-		temp->Write(1, 0, 0); temp->Write(1, 1,-1); temp->Write(1, 2, 0);
-		temp->Write(2, 0, 0); temp->Write(2, 1, 1); temp->Write(2, 2, 0);
-		kerD = temp;
-
-		// Right Up Kernel
-		temp = new mat2D<int>(3, 3);
-		temp->Write(0, 0, 0); temp->Write(0, 1, 0); temp->Write(0, 2, 1);
[…]
