+++ aletheia-resources/F5/Makefile
@@ -1,5 +1,0 @@
-default:
-	javac Embed.java
-	javac Extract.java
-clean:
-	rm -f *.class */*.class */*/*.class

[…]
