+++ INSTALL.md
@@ -25,63 +25,69 @@ After that, you can execute Aletheia with:
 ```bash
 $ ./aletheia.py 
 
-./aletheia.py <command>
-
 COMMANDS:
 
-  Automatic steganalysis:
+  Automated tools:
   - auto:      Try different steganalysis methods.
+  - dci:       Predicts a set of images using DCI evaluation.
 
   Structural LSB detectors (Statistical attacks to LSB replacement):
-  - sp:            Sample Pairs Analysis (Octave vesion).
-  - ws:            Weighted Stego Attack.
-  - triples:       Triples Attack.
   - spa:           Sample Pairs Analysis.
   - rs:            RS attack.
+  - ws:            Weighted Stego Attack.
+  - triples:       Triples Attack.
+
 
   Calibration attacks to JPEG steganography:
-  - calibration:   Calibration attack to JPEG images.
+  - calibration:   Calibration attack on F5.
 
   Feature extractors:
   - srm:           Full Spatial Rich Models.
   - srmq1:         Spatial Rich Models with fixed quantization q=1c.
   - scrmq1:        Spatial Color Rich Models with fixed quantization q=1c.
   - gfr:           JPEG steganalysis with 2D Gabor Filters.
-  - hill-maxsrm:   Selection-Channel-Aware Spatial Rich Models for HILL.
+  - dctr:          JPEG Low complexity features extracted from DCT residuals.
+
 
   Embedding simulators:
-  - lsbr-sim:             Embedding using LSB replacement simulator.
-  - lsbm-sim:             Embedding using LSB matching simulator.
-  - hugo-sim:             Embedding using HUGO simulator.
-  - wow-sim:              Embedding using WOW simulator.
-  - s-uniward-sim:        Embedding using S-UNIWARD simulator.
-  - s-uniward-color-sim:  Embedding using S-UNIWARD color simulator.
-  - j-uniward-sim:        Embedding using J-UNIWARD simulator.
-  - j-uniward-color-sim:  Embedding using J-UNIWARD color simulator.
-  - hill-sim:             Embedding using HILL simulator.
-  - ebs-sim:              Embedding using EBS simulator.
-  - ebs-color-sim:        Embedding using EBS color simulator.
-  - ued-sim:              Embedding using UED simulator.
-  - ued-color-sim:        Embedding using UED color simulator.
-  - nsf5-sim:             Embedding using nsF5 simulator.
-  - nsf5-color-sim:       Embedding using nsF5 color simulator.
-  - steghide-sim:         Embedding using Steghide simulator.
-  - steganogan-sim:       Embedding using SteganoGAN simulator.
+  - lsbr-sim:             LSB replacement simulator.
+  - lsbm-sim:             LSB matching simulator.
+  - hugo-sim:             HUGO simulator.
+  - wow-sim:              WOW simulator.
+  - s-uniward-sim:        Spatial UNIWARD simulator.
+  - s-uniward-color-sim:  Spatial UNIWARD color simulator.
+  - j-uniward-sim:        JPEG UNIWARD simulator.
+  - j-uniward-color-sim:  JPEG UNIWARD color simulator.
+  - hill-sim:             HILL simulator.
+  - hill-color-sim:       HILL color simulator.
+  - ebs-sim:              EBS simulator.
+  - ebs-color-sim:        EBS color simulator.
+  - ued-sim:              UED simulator.
+  - ued-color-sim:        UED color simulator.
+  - nsf5-sim:             nsF5 simulator.
+  - nsf5-color-sim:       nsF5 color simulator.
+  - steghide-sim:         Steghide simulator.
+  - outguess-sim:         Outguess simulator.
+  - steganogan-sim:       SteganoGAN simulator.
 
   ML-based steganalysis:
   - split-sets:            Prepare sets for training and testing.
[…]
