+++ aletheialib/brute_force.py
@@ -186,7 +186,7 @@ def f5(path, password_file):
 
     # Get the directory where the resources are installed
     dir_path = os.path.dirname(os.path.realpath(__file__))
-    dir_path = os.path.join(dir_path, os.pardir, 'aletheia-resources', 'F5')
+    dir_path = os.path.join(dir_path, os.pardir, 'aletheia-cache', 'F5')
 
     os.chdir(dir_path)
 
