+++ MANIFEST.in
@@ -1,3 +1,3 @@
 recursive-include aletheia-models/ *
 recursive-include aletheia-resources/ *
-recursive-include aletheia-octave/ *
+recursive-include aletheia-cache/ *
