+++ testing/e4s.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+./aletheia.py e4s-predict aletheia-models/e4s/e4s_srm_bossbase_hill0.40_gs.model srm sample_images/bossbase
+
+
