### 2023-07-27 887dd6e
paper: several improvements
