+++ aletheia.py
@@ -254,6 +254,9 @@ def main():
     elif sys.argv[1] == "plot-dct-histogram":
         options.tools.plot_dct_histogram()
 
+    elif sys.argv[1] == "eof-extract":
+        options.tools.eof_extract()
+
     else:
         print("Wrong command!")
 
