### 2023-08-01 e625fd5
print-metadata command
