### 2023-08-01 0770721
print-metadata command and image
