+++ CITING.md
@@ -2,18 +2,18 @@
 ### How to cite Aletheia?
 
 If you use Aletheia in a publication I would appreciate if you cite it. Please
-refer to the [version-independent Zendo DOI](https://zenodo.org/record/4655945) 
+refer to the [version-independent Zendo DOI](10.5281/zenodo.4655944) 
 or use directly the folloging BibTeX entry:
 
 ```
-@software{daniel_lerch_hostalot_2021_4655945,
-  author       = {Daniel Lerch-Hostalot},
+@software{Aletheia,
+  author       = {Lerch-Hostalot, Daniel},
   title        = {Aletheia},
-  month        = apr,
-  year         = 2021,
+  month        = Aug,
+  year         = 2023,
   publisher    = {Zenodo},
-  version      = {v0.1},
-  doi          = {10.5281/zenodo.4655945},
-  url          = {https://doi.org/10.5281/zenodo.4655945}
+  version      = {v0.2},
+  doi          = {10.5281/zenodo.4655944},
+  url          = {https://doi.org/10.5281/zenodo.4655944}
 }
 ```
