### 2023-08-08 #5653 @daniellerch
Thank you for your feedback, @danielskatz. We apologize for the delay in our response, as we have had to make several adjustments based on your observations.

1. The core code of Aletheia is written in Python. However, in the field of steganography, it's common to implement stegosystems in Matlab. That's why Aletheia also contains portions of its code in Octave. At times, this Octave code requires C implementations for optimization. It's crucial, especially for research experiments, to use the original code provided by authors. That's the primary reason for retaining the Octave code. However, we've relocated this code to an external repository, and now Aletheia downloads it only upon user request and confirmation. This process is detailed in the paper.

2. We have expanded the paper and added some illustrative examples to provide a deeper understanding.

3. We believe we have addressed the third point regarding the potentially missing DOIs.

Thank you for your patience, and we hope these changes clarify the concerns you raised. Looking forward to your further insights.
