### 2023-08-24 5c226a6
dct diffs: print rates
