### 2023-08-24 1060363
dct diffs: print rates
