+++ aletheialib/attacks.py
@@ -513,7 +513,7 @@ def print_dct_diffs(cover, stego):
         print("\nChannel "+str(i)+":")
         num_coeffs = C_jpeg.coeffs(i).shape[0]*C_jpeg.coeffs(i).shape[1]
         changes = np.sum(np.abs(C_jpeg.coeffs(i)-S_jpeg.coeffs(i)))
-        print("Changes:", changes, ", rate:", round(changes/num_coeffs,2))
+        print("Changes:", changes, ", rate:", round(changes/num_coeffs,4))
         print("H BAR    COVER      STEGO       DIFF")
         print("------------------------------------")
         for v in [-4, -3, -2, -1, 0, 1, 2, 3, 4]:
