### 2023-08-24 18b6512
structural options
