+++ octave-requirements.txt
@@ -1,1 +1,1 @@
-image signal
+image signal nan
