### 2023-09-02 0d85a86
J-MiPOD EfficientNet B0 models
