### 2023-09-12 97a4100
chang print-dct-diffs behaviour
