+++ aletheialib/attacks.py
@@ -520,7 +520,7 @@ def print_dct_diffs(cover, stego):
         for v in [-4, -3, -2, -1, 0, 1, 2, 3, 4]:
             cover = np.sum(C_jpeg.coeffs(i)==v)
             stego = np.sum(S_jpeg.coeffs(i)==v)
-            var = cover-stego
+            var = stego-cover
             print(f"{v:+}: {cover:10d} {stego:10d} {var:10d}")
 
 
