### 2023-10-02 #5653 @daniellerch
@mstimberg  I am pleased to provide a list of researches in the field, whom I believe would provide insightful feedback. I hold no personal or professional affiliations with them, and am unaware of their interest in reviewing the paper:

- Jan Butora
- Eugene Khvedchenya (BloodAxe)
- Guanshuo Xu (GuanshuoXu)
- Mehdi Boroumand
- Marc Chaumont (marcchaumont)

On the other hand, I believe @YassineYousfi would be an excellent choice, and he might also be able to recommend other suitable candidates.
