+++ aletheialib/options/ml.py
@@ -45,24 +45,22 @@ def split_sets():
     seed=int(sys.argv[7])
 
 
-    cover_files = sorted(glob.glob(os.path.join(cover_dir, '*')))
-    stego_files = sorted(glob.glob(os.path.join(stego_dir, '*')))
+    cover_files = np.array(sorted(glob.glob(os.path.join(cover_dir, '*'))))
+    stego_files = np.array(sorted(glob.glob(os.path.join(stego_dir, '*'))))
 
     if len(cover_files)!=len(stego_files):
         print("ERROR: we expect the same number of cover and stego files");
         sys.exit(0)
 
-    from sklearn.model_selection import train_test_split
-    trn_C_files, tv_C_files, trn_S_files, tv_S_files = \
-        train_test_split(cover_files, stego_files, 
-                         test_size=n_valid+n_test*2, random_state=seed)
 
-    train_C_files = trn_C_files
-    train_S_files = trn_S_files
-    valid_C_files = tv_C_files[:n_valid//2]
-    valid_S_files = tv_S_files[:n_valid//2]
-    test_C_files = tv_C_files[n_valid//2:n_valid//2+n_test//2]
-    test_S_files = tv_S_files[n_valid//2+n_test//2:n_valid//2+n_test]
+    indices = list(range(len(cover_files)))
+    random.seed(seed)
+    random.shuffle(indices)
+
+    valid_indices = indices[:n_valid//2]
+    test_C_indices = indices[n_valid//2:n_valid//2+n_test//2]
+    test_S_indices = indices[n_valid//2+n_test//2:n_valid//2+n_test]
+    train_indices = indices[n_valid//2+n_test:]
 
     train_C_dir = os.path.join(output_dir, "train", "cover")
     train_S_dir = os.path.join(output_dir, "train", "stego")
@@ -71,9 +69,9 @@ def split_sets():
     test_C_dir = os.path.join(output_dir, "test", "cover")
     test_S_dir = os.path.join(output_dir, "test", "stego")
 
+
     if not os.path.isdir(output_dir):
         os.mkdir(output_dir)
-
     os.makedirs(train_C_dir, exist_ok=True)
     os.makedirs(train_S_dir, exist_ok=True)
     os.makedirs(valid_C_dir, exist_ok=True)
@@ -81,22 +79,22 @@ def split_sets():
     os.makedirs(test_C_dir, exist_ok=True)
     os.makedirs(test_S_dir, exist_ok=True)
     
-    for f in train_C_files:
+    for f in cover_files[train_indices]:
         shutil.copy(f, train_C_dir)
 
-    for f in train_S_files:
+    for f in stego_files[train_indices]:
         shutil.copy(f, train_S_dir)
 
-    for f in valid_C_files:
+    for f in cover_files[valid_indices]:
         shutil.copy(f, valid_C_dir)
 
-    for f in valid_S_files:
+    for f in stego_files[valid_indices]:
         shutil.copy(f, valid_S_dir)
 
-    for f in test_C_files:
+    for f in cover_files[test_C_indices]:
         shutil.copy(f, test_C_dir)
 
-    for f in test_S_files:
+    for f in stego_files[test_S_indices]:
         shutil.copy(f, test_S_dir)
 # }}}
 
@@ -324,7 +322,7 @@ def create_actors():
 def effnetb0():
 
     if len(sys.argv)<7:
-        print(sys.argv[0], "effnetb0 <trn-cover-dir> <trn-stego-dir> <val-cover-dir> <val-stego-dir> <model-file> [dev] [ES]\n")
[…]
