+++ aletheialib/options/embsim.py
@@ -1,8 +1,11 @@
 import sys
 import os
+import numpy as np
+from imageio import imread, imwrite
 from aletheialib.utils import download_octave_code
 from aletheialib.utils import download_octave_aux_file
 from aletheialib.utils import download_octave_jpeg_toolbox
+from aletheialib.utils import download_F5
 
 doc="\n" \
 "  Embedding simulators:\n" \
@@ -27,6 +30,7 @@ doc="\n" \
 "  - steghide-sim:         Steghide simulator.\n" \
 "  - outguess-sim:         Outguess simulator.\n" \
 "  - steganogan-sim:       SteganoGAN simulator." 
+"  - f5-sim:               F5 simulator." 
 
 
 
@@ -371,7 +375,151 @@ def steganogan():
     sys.exit(0);
 # }}}
 
+# {{{ f5
+def f5():
 
+    if len(sys.argv)!=5:
+        print(sys.argv[0], "f5-sim <image/dir> <payload> <output-dir>")
+        sys.exit(0)
+
+    import aletheialib.stegosim
+    import aletheialib.utils
+
+    download_octave_jpeg_toolbox()
+    download_F5()
+
+    aletheialib.stegosim.embed_message(aletheialib.stegosim.f5, sys.argv[2], sys.argv[3], sys.argv[4],
+                  embed_fn_saving=True)
+    sys.exit(0);
+# }}}
+
+# {{{ adversarial_adaptive
+def adversarial_adaptive():
+
+    if len(sys.argv)<6:
+        print(sys.argv[0], "adversarial-adaptive-sim <image/dir> <payload> <output-dir> <model-file> [dev]\n")
+        sys.exit(0)
+
+
+    input_dir = sys.argv[2]
+    payload = sys.argv[3]
+    output_dir = sys.argv[4]
+    model_file=sys.argv[5]
+
+    if len(sys.argv)<7:
+        dev_id = "CPU"
+        print("'dev' not provided, using:", dev_id)
+    else:
+        dev_id = sys.argv[6]
+
+    if dev_id == "CPU":
+        print("Running with CPU. It could be very slow!")
+
+    os.environ["CUDA_VISIBLE_DEVICES"] = dev_id
+    os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
+
+    import glob
+    import aletheialib.stegosim
+    import aletheialib.models
+
+    nn = aletheialib.models.NN("effnetb0")
+    nn.load_model(model_file)
+
+    # Generate gradient files
+    if not os.path.exists("/tmp/gradients"):
+        os.mkdir("/tmp/gradients")
+
+    for f in glob.glob(os.path.join(input_dir, '*')):
+        bn = os.path.basename(f)
+        gradient = nn.get_gradient(f, 0)
+        gradient_path = os.path.join("/tmp/gradients", bn)
+        print("Save gradient", gradient_path)
+        np.save(gradient_path, gradient)
+
+    aletheialib.stegosim.embed_message(
+        aletheialib.stegosim.adversarial_adaptive, input_dir, payload, output_dir)
+
+
+    sys.exit(0);
+# }}}
+
+# {{{ adversarial_fix
+def adversarial_fix():
+
+    if len(sys.argv)<6:
+        print(sys.argv[0], "adversarial-fix <image/dir> <output-dir> <model-file> [dev]\n")
+        sys.exit(0)
+
+
+    input_dir = sys.argv[2]
+    output_dir = sys.argv[3]
+    model_file=sys.argv[4]
+
+    if len(sys.argv)<6:
+        dev_id = "CPU"
+        print("'dev' not provided, using:", dev_id)
+    else:
+        dev_id = sys.argv[5]
+
+    if dev_id == "CPU":
+        print("Running with CPU. It could be very slow!")
+
[…]
