### 2023-10-25 #5982 @mstimberg
👋🏼  @daniellerch, @YassineYousfi, @ragibson, this is the review thread for the paper. All of our communications will happen here from now on.

at the top of a new comment in this thread.

There are additional guidelines in the [first comment of this issue](https://github.com/openjournals/joss-reviews/issues/5982#issue-1960897370).

Please don't hesitate to ping me (@mstimberg) if you have any questions/concerns.
