### 2023-11-11 #5982 @ragibson
Finally getting around to this.

Just a quick note on the list of authors -- it seems the primary author @danielskatz wrote the software package and both authors collaborated on a few of the detection techniques used by the package (e.g., https://dl.acm.org/doi/10.1145/3335203.3335738 and https://ieeexplore.ieee.org/document/9722958), so the authors list looks good to me.

There are a handful of other contributors on the GitHub repository, but they're all much more minor.
