### 2023-11-11 daniellerch/aletheia#23 @ragibson
Paper description has duplicated paragraph
I was reading through the paper in https://github.com/openjournals/joss-reviews/issues/5982 and noticed this paragraph is (essentially) duplicated.

It looks like the second one is the edited version and the first one was just accidentally not removed. I.e., the section starts out with this introduction,

https://github.com/daniellerch/aletheia/blob/31409d265a2b381c8427cc0cbf3a911155513de3/paper/paper.md?plain=1#L78-L90

immediately followed by

https://github.com/daniellerch/aletheia/blob/31409d265a2b381c8427cc0cbf3a911155513de3/paper/paper.md?plain=1#L92-L103
