### 2023-11-11 daniellerch/aletheia#24 @ragibson
Minor clarifications/wordsmithing in paper's usage examples
(From https://github.com/openjournals/joss-reviews/issues/5982)

In this part of the paper, I think it would be more clear to describe the payload range of `0.05-0.25` as, for example, "ranging from 5% to 25% of the image's maximum capacity". Otherwise, it's not immediately clear what a "random 0.05 payload" refers to.

https://github.com/daniellerch/aletheia/blob/31409d265a2b381c8427cc0cbf3a911155513de3/paper/paper.md?plain=1#L108-L116

-------

Then, in this table I'd avoid abbreviating "being stego" and would prefer something like "being steganographed".

https://github.com/daniellerch/aletheia/blob/31409d265a2b381c8427cc0cbf3a911155513de3/paper/paper.md?plain=1#L132-L145

Alternatively, if that caption is in the output of the `./aletheia.py auto` command, perhaps clarify it above.

Something like "For example, below is an automated analysis of a set of images, showing the modeled probabilities of each image being generated by various steganographic methods."

https://github.com/daniellerch/aletheia/blob/31409d265a2b381c8427cc0cbf3a911155513de3/paper/paper.md?plain=1#L125-L127
