### 2023-11-11 79bef25
Update paper.md
Remove duplicated paragraph.
