### 2023-11-11 daniellerch/aletheia#23 @daniellerch
The duplicated paragraph was indeed a mistake, where the first version wasn't removed after revisions. It's fixed now.
Thank you for the notification.
