### 2023-11-11 #5982 @ragibson
A note on the data sharing point -- the paper basically contains no original data since the examples are generic and I was able to effectively run them with my own local images.
