### 2023-11-11 #5982 @ragibson
I would note that I ran into a few opaque errors in the simulators, but they look like research-grade code sourced from other authors/institutions (hence the requirement to accept secondary licenses and download the code from https://github.com/daniellerch/aletheia-external-resources). In these cases, the issues are external to the authors' work.

@daniellerch That said, I would like to see the secondary repository explicitly linked in the paper and/or READMEs of the package rather than just in `aletheialib/utils.py`.
