### 2023-11-11 daniellerch/aletheia#26 @ragibson
Many newer commands are not included in the documentation
(From https://github.com/openjournals/joss-reviews/issues/5982)

I noticed this when trying to find documentation on the paper's usage of `hill-color-sim`. It seems like the discussion in https://daniellerch.me/stego/aletheia/intro-en/ is somewhat out of date as a lot of the commands have been changed, removed, or added.

To name a few, `j-mipod-sim`, `hill-color-sim`, `outguess-sim`, `actor-predict-fea`, `brute-force-f5`, `eof-extrat` are printed by `aletheia.py`, but are not included in that documentation page. It seems to affect most of the sections, so an update would be greatly appreciated.
