### 2023-11-11 daniellerch/aletheia#27 @ragibson
Frequent false positives in `aletheia.py auto`?
(From https://github.com/openjournals/joss-reviews/issues/5982)

I tried to run an example of the paper/documentation's mention of "the auto command, which performs an exploratory analysis trying to identify the steganalysis technique used."

Is that functionality only meant to be used for images *known* to contain covert information? If so, that should be clearly listed in the documentation since it does not seem accurate in determining *if* an image contains a covert message.

On a handful of images (from a phone), a screenshot, and a 512x512 image of blank white square, the analysis seems to indicate all of them have a high likelihood of being steganographed.
```bash
$ ./aletheia.py auto example_images/
                     Outguess  Steghide   nsF5  J-UNIWARD *
-----------------------------------------------------------
20231110_3.jpg         [1.0]     0.0     [0.9]    [0.8]
white_square_512.jpg    0.1      0.1     [0.6]    [0.8]
20231110_1.jpg         [1.0]    [1.0]    [1.0]    [0.8]
20231110_2.jpg         [1.0]     0.0     [0.7]    [0.7]
20231111_4.jpg         [1.0]     0.0     [1.0]    [0.9]
screenshot_202311...    0.0      0.0     [0.6]     0.2

* Probability of being stego using the indicated steganographic method.
```
However, none of these images have been steganographically altered in any way. In fact, `white_square_512.jpg` is completely white.
