### 2023-11-11 daniellerch/aletheia#28 @ragibson
Unattributed password list / license conflict in `aletheia-resources/`?
(From https://github.com/openjournals/joss-reviews/issues/5982)

The password list `aletheia-resources/passwords.txt` appears to be taken from the OWASP SecLists project, which is either under CC-by-SA 3.0 (mentioned in https://en.wikipedia.org/wiki/Wikipedia:10,000_most_common_passwords) or the MIT License [(2018 Daniel Miessler)](https://github.com/danielmiessler/SecLists/blob/master/LICENSE).

That said, the version in this repository has a very small number of edits where two of the passwords have been moved around and an erroneous "j" has been added on line 952663. Regardless, attribution is required and/or it should be moved to https://github.com/daniellerch/aletheia-external-resources, though I'm not sure if this file is needed by any of the scripts.

E.g.,
```diff
@@ -506891,7 +506891,6 @@ legally
 Legalize
 legalise
 legal123
-lega
 leg522
 leg0las
 leftys
@@ -879413,6 +879412,7 @@ xmanjadas
 xmanifestox
 xmango12
 xmanfan
+lega
 xmandrad
 xmandc
 xmandate
```
and
```diff
@@ -952660,6 +952660,7 @@ WddFKa
 Wddcbri2
 wdday1
 WDDaXuqvWXcJ
+j
 WdDAukDK
 WDd96Q
 wdd7rezl
```
