### 2023-11-11 daniellerch/aletheia#27 @ragibson
This is discussed more in several of the practical attack articles -- it turns out the models do not generalize particularly well and are only really reliable on images similar to the training set. C.f. the discussions of the `dci` command in https://daniellerch.me/stego/aletheia/steghide-attack-en/ and https://daniellerch.me/stego/aletheia/f5-attack-en/.

I still believe this should be more clearly called out beforehand when discussing the `auto` command, especially since it is introduced as a tool for an initial exploratory analysis.
