### 2023-11-11 daniellerch/aletheia#29 @ragibson
Test scripts should be documented (and fixed) for verifying successful installation
(From https://github.com/openjournals/joss-reviews/issues/5982)

There are four scripts in `testing/`, but they don't seem to be documented. Do they return a nonzero error code if they fail? Which commands do each of them test?

Ideally, these could be called by a script/command to verify that the installation is successful or manual steps to run them should be documented with the expected output. Running them on a GitHub action for new commits would probably also be a good idea.

My assumption is below.

## testing/e4s.sh

Appears to be functional and produces
```
241.png Stego
5300_hill.png Stego
349.png Cover
151.png Stego
9966_hill.png Stego
7731_hill.png Stego
18.png Stego
5003.png Stego
8828.png Stego
321_hill.png Stego
712.png Stego
3186.png Stego
6166_hill.png Stego
1.png Stego
1021_hill.png Stego
2052_hill.png Cover
4119_hill.png Stego
2003_hill.png Cover
3178_hill.png Stego
1027.png Stego
```

## testing/feaext.sh

Appears to be functional and produces
```
Testing gfr / sample_images/alaska2jpg ...
JPEG quality not provided, using detection via 'identify'
Number of rotations for Gabor kernel no provided, using: 32
15 fea.txt
Testing dctr / sample_images/alaska2jpg ...
JPEG quality not provided, using detection via 'identify'
15 fea.txt
Testing srm / sample_images/alaska2 ...
12 fea.txt
Testing srmq1 / sample_images/lena_gs.png ...
1 fea.txt
Testing scrmq1 / sample_images/alaska2 ...
12 fea.txt
```

## testing/lsbdetect.sh

Appears to be functional and produces
```
Testing spa / sample_images/lena_lsbr.png ...
Hidden data found in channel R 0.09308090623358549
Hidden data found in channel G 0.09238585295279302
Hidden data found in channel B 0.11546638236749293
Testing rs / sample_images/lena_lsbr.png ...
Hidden data found in channel R 0.19283881618132384
Hidden data found in channel G 0.18940002370415496
Hidden data found in channel B 0.19726342116628495
Testing ws / sample_images/lena_lsbr.png ...
Hidden data found in channel R 0.10590840834668327
Hidden data found in channel G 0.07463418193363092
Hidden data found in channel B 0.07968467118722876
Testing triples / sample_images/lena_lsbr.png ...
Hidden data found in channel R 0.09769010200708361
Hidden data found in channel G 0.06694333522398271
```

## testing/stegosim.sh

This script crashes, seemingly due to steganogan incompatibilities with pytorch. It prints the first few tests seemingly fine, but the next one fails (perhaps due to an upstream error / incompatibility with recent pytorch changes -- see https://github.com/DAI-Lab/SteganoGAN/issues/90).
```
Testing lsbr-sim / sample_images/lena.png ...
Using 8 threads
Testing lsbm-sim / sample_images/lena.png ...
Using 8 threads
Testing wow-sim / sample_images/lena_gs.png ...
Using 8 threads
Testing s-uniward-sim / sample_images/lena_gs.png ...
Using 8 threads
Testing s-uniward-color-sim / sample_images/lena.png ...
Using 8 threads
Testing j-uniward-sim / sample_images/lena_gs.jpg ...
Using 8 threads
Testing j-uniward-color-sim / sample_images/lena.jpg ...
Using 8 threads
Testing hill-sim / sample_images/lena_gs.png ...
Using 8 threads
Testing hill-color-sim / sample_images/lena.png ...
Using 8 threads
Testing ebs-sim / sample_images/lena_gs.jpg ...
Using 8 threads
Testing ebs-color-sim / sample_images/lena.jpg ...
Using 8 threads
Testing ued-sim / sample_images/lena_gs.jpg ...
Using 8 threads
Testing ued-color-sim / sample_images/lena.jpg ...
Using 8 threads
Testing nsf5-sim / sample_images/lena_gs.jpg ...
Using 8 threads
Testing nsf5-color-sim / sample_images/lena.jpg ...
Using 8 threads
Testing steghide-sim / sample_images/lena.jpg ...
Using 8 threads
Testing outguess-sim / sample_images/lena.jpg ...
Using 8 threads
Testing steganogan-sim / sample_images/lena.png ...
Using 8 threads
~/.local/lib/python3.10/site-packages/torch/serialization.py:1101: SourceChangeWarning: source code of class 'torch.nn.modules.container.Sequential' has changed. you can retrieve the original source code by accessing the object's source attribute or set `torch.nn.Module.dump_patches = True` and use the patch tool to revert the changes.
  warnings.warn(msg, SourceChangeWarning)
~/.local/lib/python3.10/site-packages/torch/serialization.py:1101: SourceChangeWarning: source code of class 'torch.nn.modules.conv.Conv2d' has changed. you can retrieve the original source code by accessing the object's source attribute or set `torch.nn.Module.dump_patches = True` and use the patch tool to revert the changes.
  warnings.warn(msg, SourceChangeWarning)
~/.local/lib/python3.10/site-packages/torch/serialization.py:1101: SourceChangeWarning: source code of class 'torch.nn.modules.activation.LeakyReLU' has changed. you can retrieve the original source code by accessing the object's source attribute or set `torch.nn.Module.dump_patches = True` and use the patch tool to revert the changes.
  warnings.warn(msg, SourceChangeWarning)
~/.local/lib/python3.10/site-packages/torch/serialization.py:1101: SourceChangeWarning: source code of class 'torch.nn.modules.batchnorm.BatchNorm2d' has changed. you can retrieve the original source code by accessing the object's source attribute or set `torch.nn.Module.dump_patches = True` and use the patch tool to revert the changes.
  warnings.warn(msg, SourceChangeWarning)
~/.local/lib/python3.10/site-packages/torch/serialization.py:1101: SourceChangeWarning: source code of class 'steganogan.decoders.DenseDecoder' has changed. you can retrieve the original source code by accessing the object's source attribute or set `torch.nn.Module.dump_patches = True` and use the patch tool to revert the changes.
  warnings.warn(msg, SourceChangeWarning)
~/.local/lib/python3.10/site-packages/torch/serialization.py:1101: SourceChangeWarning: source code of class 'steganogan.critics.BasicCritic' has changed. you can retrieve the original source code by accessing the object's source attribute or set `torch.nn.Module.dump_patches = True` and use the patch tool to revert the changes.
  warnings.warn(msg, SourceChangeWarning)
Traceback (most recent call last):
  File "~/.local/bin/steganogan", line 8, in <module>
    sys.exit(main())
  File "~/.local/lib/python3.10/site-packages/steganogan/cli.py", line 88, in main
    args.action(args)
  File "~/.local/lib/python3.10/site-packages/steganogan/cli.py", line 25, in _encode
    steganogan = _get_steganogan(args)
  File "~/.local/lib/python3.10/site-packages/steganogan/cli.py", line 20, in _get_steganogan
    return SteganoGAN.load(**steganogan_kwargs)
  File "~/.local/lib/python3.10/site-packages/steganogan/models.py", line 364, in load
    steganogan = torch.load(path, map_location='cpu')
  File "~/.local/lib/python3.10/site-packages/torch/serialization.py", line 1028, in load
    return _legacy_load(opened_file, map_location, pickle_module, **pickle_load_args)
  File "~/.local/lib/python3.10/site-packages/torch/serialization.py", line 1256, in _legacy_load
    result = unpickler.load()
  File "~/.local/lib/python3.10/site-packages/torch/optim/adam.py", line 66, in __setstate__
    super().__setstate__(state)
  File "~/.local/lib/python3.10/site-packages/torch/optim/optimizer.py", line 295, in __setstate__
    self.defaults.setdefault('differentiable', False)
AttributeError: 'Adam' object has no attribute 'defaults'
```
