### 2023-11-15 daniellerch/aletheia#24 @daniellerch
Thank you for your suggestion. I have implemented the suggested changes.
