### 2023-11-15 daniellerch/aletheia#29 @daniellerch
I can't reproduce the error, but I think it might be related to the TensorFlow version. I am using TensorFlow 2.0.0. Could you please specify which version of TensorFlow you are using?
