### 2023-11-15 daniellerch/aletheia#28 @daniellerch
I don't recall exactly where I obtained the original password file, so I've replaced it with a similar one from Daniel Miessler, and have also included the license.

On another note, the scripts don't use this file; I'm providing it merely as a convenience for users who want to utilize the brute force attacks offered by Aletheia without having to search for a password file themselves.
