### 2023-11-16 daniellerch/aletheia#29 @ragibson
I am using TensorFlow 2.15.0.
```bash
$ pip freeze | grep -E "stegan|tensorflow"
steganogan==0.1.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
```
EDIT: I would also note that TensorFlow 2.0.0 is >4 years old
