### 2023-11-19 daniellerch/aletheia#30 @YassineYousfi
incorrect use of word "bitmaps"
(From https://github.com/openjournals/joss-reviews/issues/5982)

In https://github.com/daniellerch/aletheia/blob/master/README.md/ and https://github.com/daniellerch/aletheia/blob/master/aletheia-models/README.md
I'm assuming this means uncompressed images?
