### 2023-11-19 daniellerch/aletheia#31 @YassineYousfi
`auto` models do not support grayscale images 
(From https://github.com/openjournals/joss-reviews/issues/5982)

Wanted to try this nice stego image embedded with HILL (and STCs) but the tool raised this error.
```
WARNING: image ignored: /home/batman/aletheia/sample_images/test_yassine/157.png , expected number of channels: 3
ERROR: please provide valid files
```
This limitation should be outlined somewhere.
I do see some models trained on grayscale images here though : https://github.com/daniellerch/aletheia/blob/master/aletheia-models/README.md#grayscale-models
![157](https://github.com/daniellerch/aletheia/assets/29985433/fd607f30-ad53-453c-aa65-68b27fa6ab18)
