### 2023-11-19 #5982 @YassineYousfi
Some notes to improve the submission:

- The target audience for the software is not clearly stated in the documentation. I am assuming this software is geared towards researchers and enthusiasts.

- How were the machine learning models trained? Very little detail is provided in https://github.com/daniellerch/aletheia/blob/master/aletheia-models/README.md.
Note that some open source steganalysis training scripts exist (https://github.com/DDELab/deepsteganalysis), but weren't cited. can they be used to contribute more or better models?
