### 2023-11-19 daniellerch/aletheia#32 @YassineYousfi
detection thresholds should not be hard coded
(From https://github.com/openjournals/joss-reviews/issues/5982)

in https://github.com/daniellerch/aletheia/blob/master/aletheialib/options/structural.py#L43 for example, if the user wants to detect a payload lower than 0.05 bpp.
