### 2023-11-22 daniellerch/aletheia#26 @daniellerch
Thank you for pointing out the inconsistencies in the documentation. I have now updated it to include the new commands.
