### 2023-11-22 daniellerch/aletheia#30 @daniellerch
Yes, the term "bitmap" in the documentation indeed refers to uncompressed images. To avoid any confusion, I have updated the documentation to clearly reflect this. Your input is greatly appreciated.
