### 2023-11-29 a829765
Optional thresholds. Issue #32
