### 2023-11-29 daniellerch/aletheia#32 @daniellerch
I have implemented the change. Now threshold the threshold can be specified by the user.
Thank you for the note.
