### 2023-11-29 daniellerch/aletheia#31 @daniellerch
Thank you for raising this issue. I wanted to inform you that the older grayscale image models are no longer maintained.
This change has been documented for clarity:
https://daniellerch.me/stego/aletheia/intro-en/#automated-tools
