### 2023-11-29 daniellerch/aletheia#27 @daniellerch
I have updated the documentation to clarify that Aletheia models may be susceptible to CSM issues. To determine if the models are reliable for the images you're attempting to analyze, using the 'dci' command is essential. Additionally, I've expanded the documentation to include the 'dci' command in the introduction, although it was already mentioned in other sections.

You can find the changes here:
https://daniellerch.me/stego/aletheia/intro-en/#automated-tools

Thank you for your comments.
