### 2023-11-29 daniellerch/aletheia#29 @daniellerch
My apologies, I hadn't read the log you sent me thoroughly. The issue seems to be with SteganoGAN and its use of the Torch library. This is not related to Aletheia, so it's not something I can resolve. I have removed SteganoGAN from the test script.
Also, please note that these scripts are not designed for user execution to verify installations. They are tools for my use, to ensure everything continues to function correctly after code changes.
