+++ EXTERNAL.md
@@ -1,0 +1,17 @@
+
+### External resources
+
+Aletheia utilizes external software sourced from various repositories and 
+projects. Due to license incompatibilities with Aletheia's MIT license, these 
+components are not directly included in our main repository. As a result, we 
+have moved all such software to an external repository 
+'[aletheia-external-resources](https://github.com/daniellerch/aletheia-external-resources)'. 
+
+Whenever any of these components are used, the corresponding license is presented 
+to the user for review and acceptance. This approach ensures adherence to 
+licensing regulations and maintains transparency in the use of third-party 
+software. It allows us to efficiently integrate diverse tools and technologies 
+while respecting the copyrights and legal conditions set by their creators, 
+all within the framework of Aletheia's MIT license compatibility
+
+
