### 2023-12-12 #5982 @daniellerch
@YassineYousfi, thank you for your feedback. I have updated the README to better define the target audience. In addition, I've added more details about the training of Aletheia's machine learning models; there's a comprehensive article on this topic. Regarding DDELab's software, while it isn't directly integrated into Aletheia due to license incompatibilities, it is utilized by Aletheia and is available in an external repository. You can find the citation and acknowledgments there. Please refer to the 'External resources' section in the README.
