### 2023-12-14 #5982 @ragibson
Checking off a few more items from my checklist:
- **Functionality documentation**: I think this is good -- it's a bit of a strange case because a lot of the underlying tools are work from other authors, so they're not directly documented.
- **Automated tests**: These do not exist since the [testing scripts are apparently not meant for general use](https://github.com/daniellerch/aletheia/issues/29#issuecomment-1831960788), but I'm willing to accept the program's help text on run and documentation examples as roughly verifying "the functionality of the software".
- **References:** I have not thoroughly checked, but I'm not sure if all the open-source steganography/steganalysis tools are cited. However, their licenses are printed out by Aletheia itself when each one is downloaded for the first time, so I think this is okay.

Otherwise,
- **Community guidelines**: As far as I can tell, guidelines on contributing/reporting issues/etc. do not exist.
- **State of the field**: The paper does mention that Aletheia "stands out as the sole steganalysis tool currently available that incorporates the latest detection techniques." Does it need to / should it explicitly mention other commonly-used similar packages?
- I know @YassineYousfi had some other papers that he wanted to see potentially cited.
