### 2024-01-10 #5982 @mstimberg
:wave: Hi everyone, and a Happy New Year :sparkles: ! I hope you are all doing well and had a good break (assuming you took one, of course).

@daniellerch could you give us an update with respect to the remaining issues in @ragibson's comment above?
@YassineYousfi do the changes mentioned in @daniellerch's [earlier comment](https://github.com/openjournals/joss-reviews/issues/5982#issuecomment-1852567888) address all your concerns?
