+++ CONTRIBUTING.md
@@ -1,0 +1,14 @@
+
+### How-to contribute
+We appreciate your interest in contributing to [Aletheia](https://github.com/daniellerch/aletheia). 
+As an open-source project, we welcome enhancements and additional features.
+
+## Report issues or problems with the software
+
+For inquiries, reporting issues, or suggesting improvements, please submit 
+them as an [issue on Github](https://github.com/daniellerch/aletheia/issues/new). 
+Although direct user support isn't guaranteed, we aim to tackle these issues as 
+they help refine our documentation and codebase. Suggestions for new features 
+are also encouraged.
+
+
