+++ CONTRIBUTING.md
@@ -3,7 +3,7 @@
 We appreciate your interest in contributing to [Aletheia](https://github.com/daniellerch/aletheia). 
 As an open-source project, we welcome enhancements and additional features.
 
-## Report issues or problems with the software
+### Report issues or problems with the software
 
 For inquiries, reporting issues, or suggesting improvements, please submit 
 them as an [issue on Github](https://github.com/daniellerch/aletheia/issues/new). 
