### 2024-01-12 #5982 @mstimberg
Thank you very much @YassineYousfi @daniellerch for getting back to me, and thanks for your reviews in the first place, obviously.

@daniellerch since both reviewers have recommended the paper for acceptance, I will now proceed with the final checks before the final publishing steps. Could you please take care of the author tasks in the checklist below :point_down:
