### 2024-01-12 ba88b03
Fix capitalization in references
