+++ paper/paper.bib
@@ -1,12 +1,12 @@
 
 @ARTICLE{Boroumand:2019:SRNet, 
     author={M. {Boroumand} and M. {Chen} and J. {Fridrich}}, 
-    journal={{IEEE} {T}ransactions on {I}nformation {F}orensics and {S}ecurity}, 
-    title={{D}eep {R}esidual {N}etwork for {S}teganalysis of {D}igital {I}mages}, 
+    journal={{IEEE} Transactions on Information Forensics and Security},
+    title={Deep Residual Network for Steganalysis of Digital Images},
     year={2019}, 
     volume={14}, 
     number={5}, 
-    pages={1181-1193}, 
+    pages={1181--1193},
     month={May},
     doi={10.1109/TIFS.2018.2871749}
 }
@@ -19,7 +19,7 @@ isbn = {9781595938572},
 publisher = {Association for Computing Machinery},
 address = {New York, NY, USA},
 booktitle = {Proceedings of the 9th Workshop on Multimedia \& Security},
-pages = {3–14},
+pages = {3–-14},
 numpages = {12},
 location = {Dallas, Texas, USA},
 series = {MM\&Sec '07}
@@ -27,12 +27,12 @@ series = {MM\&Sec '07}
 
 @ARTICLE{Guo:2014:UED, 
     author={L. {Guo} and J. {Ni} and Y. Q. {Shi}}, 
-    journal={{IEEE} {T}ransactions on {I}nformation {F}orensics and {S}ecurity}, 
-    title={{U}niform {E}mbedding for {E}fficient {JPEG} {S}teganography}, 
+    journal={{IEEE} Transactions on Information Forensics and Security},
+    title={Uniform Embedding for Efficient JPEG Steganography},
     year={2014}, 
     volume={9}, 
     number={5}, 
-    pages={814-825}, 
+    pages={814--825},
     month={May},
 }
 
@@ -56,7 +56,7 @@ doi={10.1007/11552055_12}
     journal={{EURASIP} {J}ournal on {I}nformation {S}ecurity},
     volume={2014},
     number={1},
-    pages={1-13},
+    pages={1--13},
     year={2014},
     publisher={Springer-Verlag},
     doi={10.1186/1687-417X-2014-1}
@@ -64,32 +64,13 @@ doi={10.1007/11552055_12}
 
 @inproceedings{Ker:2013:real_world,
     author = {Ker, A. D. and Bas, P. and B\"{o}hme, R. and Cogranne, R. and Craver, S. and Filler, T. and Fridrich, J. and Pevn\'{y}, T.},
-    title = {{M}oving {S}teganography and {S}teganalysis from the {L}aboratory into the {R}eal {W}orld},
-    booktitle = {{P}roceedings of the {F}irst {ACM} {W}orkshop on {I}nformation {H}iding and {M}ultimedia {S}ecurity},
+    title = {Moving Steganography and Steganalysis from the Laboratory into the Real World},
+    booktitle = {Proceedings of the First ACM Workshop on Information Hiding and Multimedia Security},
     year = {2013},
-    pages = {45-58},
+    pages = {45--58},
     publisher = {Association for Computing Machinery},
     address = {New York, NY, USA},
 } 
-
-
-@software{Lerch-Hostalot:Aletheia,
-   author = {Lerch-Hostalot, Daniel},
-   doi = {10.5281/zenodo.4655945},
-   month = apr,
-   title = {{Aletheia}},
-   version = {v0.1},
-   year = {2021}
-}
-
-@misc{Lerch-Hostalot:Aletheia,
-   author = {Lerch-Hostalot, Daniel},
-   title = {Aletheia},
-   year = {2023},
-   publisher = {GitHub},
-   journal = {GitHub repository},
[…]
