+++ paper/paper.bib
@@ -22,7 +22,8 @@ booktitle = {Proceedings of the 9th Workshop on Multimedia \& Security},
 pages = {3–-14},
 numpages = {12},
 location = {Dallas, Texas, USA},
-series = {MM\&Sec '07}
+series = {MM\&Sec '07},
+doi = {10.1145/1288869.1288872}
 }
 
 @ARTICLE{Guo:2014:UED, 
@@ -34,6 +35,7 @@ series = {MM\&Sec '07}
     number={5}, 
     pages={814--825},
     month={May},
+    doi={10.1109/TIFS.2014.2312817}
 }
 
 @InProceedings{Hetzl:2005:steghide,
@@ -70,6 +72,7 @@ doi={10.1007/11552055_12}
     pages = {45--58},
     publisher = {Association for Computing Machinery},
     address = {New York, NY, USA},
+    doi={10.1145/2482513.2482965}
 } 
 
 @inproceedings{Lerch-Hostalot:2019,
