### 2024-01-12 4da6b8b
Change code language from bash to verb to avoid weird highlighting
