### 2024-01-12 #5982 @mstimberg
Hi again @daniellerch, in addition to the points mentioned in the author checklist above, please have a look at the PR I just opened (daniellerch/aletheia#33). I made a few minor edits (mostly related to the formatting of the references and missing DOIs). Please merge it if you agree with all changes :pray:
