### 2024-01-12 #5982 @mstimberg
Thanks for the updates @daniellerch. Regarding the Zenodo archive for v0.3, could you add the second author as an author there as well? If I understand correctly, they did not commit code directly, but supervised the development? In general, we prefer to have the same list of authors for the JOSS paper and the Zenodo archive, but if you have a strong reason to *not* have them be the same, please let me know!
Also, please:
* add the ORCID(s) to the Zenodo archive
* add the license to the Zenodo archive
* use the full title of the article as the title on Zenodo

**Note that you can make all these changes manually on Zenodo, without releasing a new version**

I think you used the automatic GitHub-Zenodo integration – if you want to make future releases have the complete metadata automatically, you can use a `.zenodo.json` file: https://developers.zenodo.org/#harvesting-with-multiple-filters
