### 2024-01-12 084123f
Update CITING.md
