### 2024-01-12 #5982 @daniellerch
@mstimberg The second author didn't directly contribute code, yet their involvement has been recognized, and they are now included in the Zenodo archive. I believe everything is now in order. Thank you!
