### 2024-01-12 #5982 @mstimberg
@daniellerch Many thanks for the changes. From your changes on Zenodo, I just realized that the affiliations concern three separate institutions (even though I assume they are all co-located). Could you update the paper in that regard, i.e. list them as ¹²³ as in the Zenodo archive (stating "Barcelona, Spain" each time – assuming this is correct, of course)? Thanks!
