### 2024-01-12 #5982 @daniellerch
@mstimberg  Thank you for pointing this out. The changes have been made to the paper, listing the affiliations as ¹²³ and stating 'Barcelona, Spain' for each, as per your suggestion. Thanks again!
