### 2024-01-12 #5982 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
