### 2024-01-15 #5982 @danielskatz
👋 @daniellerch - I've suggested some minor changes in the paper text in https://github.com/daniellerch/aletheia/pull/34 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication
