### 2024-01-16 #5982 @danielskatz
Congratulations to @daniellerch (Daniel Lerch-Hostalot) and co-author on your publication!!

JOSS depends on volunteers and couldn't be successful without you
