### 2024-01-16 #5982 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
