### 2024-01-16 #5982 @daniellerch
Thanks @YassineYousfi, @ragibson, @mstimberg and @danielskatz for your great work on the JOSS publication. Really appreciate your insightful reviews and edits!
